+++ CHANGELOG.md
@@ -27,7 +27,7 @@ Release: dd.mm.yyyy
 #### Modal
 
 - New: Use `weight`-property to adjust the weight for the modal (in points), by default `24` ([#126](https://github.com/kasper/phoenix/issues/126)).
-- New: Use `appearance`-property to set the appearance of the modal (`dark|transparent`), by default `dark` ([#126](https://github.com/kasper/phoenix/issues/126)).
+- New: Use `appearance`-property to set the appearance of the modal (`dark|light|transparent`), by default `dark` ([#126](https://github.com/kasper/phoenix/issues/126)).
 - Change: Origin can be adjusted dynamically ([#126](https://github.com/kasper/phoenix/issues/126)).
 
 #### Screen